ble: prepare DRAM for running in warm boot mode
The DRAM is marked as non-executable memory segment
in BootROM.
In the Warm Boot mode the control should be passed from
BLE to the ATF that already exists in the DRAM.
This is needed for eliminating the unnecessary boot
image reload from the boot device.
This patch marks the DRAM as an executable memory.

Change-Id: I29a9d4e6a47cd664ce7040c270537690f1f1254e
Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/40474
Tested-by: iSoC Platform CI <ykjenk@marvell.com>