fix: pm: a3700: skip setting gpio wake up bit for PPD and stepdown mode
The GPIO wake up enable registers(0xD0013C18 and 0xD0013C1C) are used
to enable the GPIO pin to wake up the CPU from partial powerdown and
step down mode. According to the suspend to RAM testing, the GPIO
wake up do not need to set it, and there is crash if some of its bits
are set.
This patch removes the setting for GPIO wake up bits.

Change-Id: Iae5ab2aac1fe5b8a4023158d4a4fe763c2723f4c
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/38683
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hua Jing <jinghua@marvell.com>