atf: ARO: Introduce Adaptive Ring Oscillator
is a clock generator, which is used in AP806 as a source
clock for the CPU clusters.
In AP806 the ARO is not the major clock, and this document discuss
the way to switch from PLL mode (which is the default mode) to ARO mode.
The motivation for ARO usage is improving the Yield of the device
at 2Ghz and improving the power consumption at that work point.
In general – ARO allow lower Vmin for running in this premium work point
which serves both cause (yield and power).

Change-Id: I76a391a29cd98abedcd003f52d37a60459ee8b80
Signed-off-by: Christine Gharzuzi <chrisg@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/40640
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>