plat: marvell: a8k: add support for scp_bl2 concatenated image
The Armada 8K has 5 service CPUs and Armada 7K has 3. Therefore it was required
to provide a method for loading firmware to all of the service CPUs. To achieve
that, the scp_bl2 image in fact is file containing up to 5 concatenated
firmwares.

This patch allows to recognize new format of scp_bl2 image, interpret it
correctly, restore built-in images and load particular image to appropriate
service CPU pointed in image header.

The new scp_bl2 file format is:
0x0:	File header (magic, nr_of_images)
0x8:	Image header 0 (type, length, version)
0x14:	Image header 1 (type, length, version)
...
0xyy	Build-in image described by image header 0
0xzz	Build-in image described by image header 1

The possible images that can be built in scp_bl2 image are firmwares for:
mss ap cm3, mss cp0 cm3, mss cp1 cm3, mg cp0 cm3 and mg cp1 cm3. Thanks to the
file format and headers for each file, ATF platform code is able to recognise
how many files are built-in and what is their destination (to which co-processor
they need to be loaded).

Change-Id: I07719eb10f43a028ebea4713ccf149b86db81ec4
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/42635
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>
Tested-by: Kostya Porotchkin <kostap@marvell.com>