pm: a8k: add generic system power off support
A generic data structure is defined to implement the system power off.
Currently only power off by GPIO is supported.
Multiple GPIOs are supported to notify the PMIC to power off the system,
also the sequence and delay can be configured.
The GPIO power off flow is divided into prepare and trigger stages, because
need to issue DDR self-refresh between them.
After DDR self-refresh is enabled, no DRAM access is allowed, thus it is
not allowed to access DRAM in GPIO trigger routine.

Change-Id: I72066dff362030fef47c0e7ad645282bd3ede026
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/43047
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>