plat/marvell: a80x0_mcbin: configure Direct Mode window for SPI flash
In order to be able to support UEFI variables, direct mode access
to the SPI flash is required. The variables will be placed at
4MB - 256KB offset. For this purpose at offset 0xf9000000 configure
new window pointing to the CP1 SPI1 CS0 device. Because the platform
supports maximum amount of 6 IO windows, CP1 PCIE0 window had to be
be extended to cover the SPI area as well.

Change-Id: I125ebfb5259549e2ea0d0d2c76296e4167c6ca45
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/46711
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>